fix: rename to HomePage & delet test-files
