+++ src/index.tsx
@@ -3,49 +3,33 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import {
-  createBrowserRouter,
-  RouterProvider,
-  Routes,
-  Route,
-} from "react-router-dom";
-import { StartWindow } from "./components/StartWindow";
-import { GamePage } from "./components/TestGamePage";
-import { BrowserRouter } from 'react-router-dom';
-/*
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { HomePage } from "./Pages/HomePage";
+import { GamePage } from "./Pages/GamePage";
+
 const router = createBrowserRouter([
   {
     path: "/",
     element: <App />,
     children: [
-      { 
-        path: "/", 
-      element: <StartWindow /> },
+      {
+        path: "/",
+        element: <HomePage />,
+      },
       {
         path: "/game",
         element: <GamePage />,
       },
-     
-     
     ],
   },
-]);*/
+]);
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
-      <BrowserRouter>
-      <Routes>
-      <Route path="/" element={<App />}>
-        <Route index element={<StartWindow/>} />
-        <Route path="game" element={<GamePage />} />
-      </Route>
-    </Routes>
-      </BrowserRouter>
-    {/* <RouterProvider router={router} />*/}
- 
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
